requirements.txt
fastapi[all]
python-dotenv
openpyxl